refactor(cmd): de-duplicate the IEC byte formatter
formatBytes (capture.go) and humanBytes (inspect.go) were byte-for-byte
identical. Consolidate into a single formatBytes in cmd/format.go, drop
both copies, and point the inspect output at it. Test collapses to the
one function.
