+++ cmd/capture.go
@@ -176,17 +176,3 @@ func startSpinner(w *os.File) func() {
 		<-done
 	}
 }
-
-// formatBytes returns a human-readable byte size string.
-func formatBytes(b int64) string {
-	const unit = 1024
-	if b < unit {
-		return fmt.Sprintf("%d B", b)
-	}
-	div, exp := int64(unit), 0
-	for n := b / unit; n >= unit; n /= unit {
-		div *= unit
-		exp++
-	}
-	return fmt.Sprintf("%.1f %ciB", float64(b)/float64(div), "KMGTPE"[exp])
-}
