+++ cmd/format.go
@@ -1,0 +1,17 @@
+package cmd
+
+import "fmt"
+
+// formatBytes returns a human-readable IEC byte size string (e.g. "1.5 KiB").
+func formatBytes(b int64) string {
+	const unit = 1024
+	if b < unit {
+		return fmt.Sprintf("%d B", b)
+	}
+	div, exp := int64(unit), 0
+	for n := b / unit; n >= unit; n /= unit {
+		div *= unit
+		exp++
+	}
+	return fmt.Sprintf("%.1f %ciB", float64(b)/float64(div), "KMGTPE"[exp])
+}
